Move app state to its own module
